+++ controllers/userController.js
@@ -5,8 +5,9 @@ const axios = require('axios')
 const Filter = require('bad-words'),
   filter = new Filter()
 filter.addWords('dicks', 'fuckton', 'fuckload', 'assload')
-const selectTemplate = require('./insultTemplates')
 let newURL = ''
+
+let randNumb = 0
 
 //get all users
 router.get('/', (req, res) => {
@@ -36,7 +37,6 @@ router.put('/update-user/:id', (req, res) => {
   User.findByIdAndUpdate(req.params.id, req.body, { new: true }).then(
     updatedUser => {
       res.json(updatedUser)
-
     }
   )
 })
@@ -79,5 +79,28 @@ router.get('/get-insult', (req, res) => {
       console.log(error)
     })
 })
+
+//start insult templates
+const selectTemplate = insultURL => {
+  randNumb = Math.floor(Math.random() * Math.floor(5))
+  console.log(`Template #${randNumb} selected!`)
+  switch (randNumb) {
+    //Sports Team insults
+    case 0:
+      return `https://insult.mattbas.org/api/insult.json?template=The+${sport}+are+as+%3Cadjective%3E+as+%3Carticle+target%3Dadj1%3E+%3Cadjective+min%3D1+max%3D3+id%3Dadj1%3E+%3Camount%3E+of+%3Cadjective+min%3D1+max%3D3%3E+%3Canimal%3E+%3Canimal_part%3E`
+    case 1:
+      return `https://insult.mattbas.org/api/insult.json?template=Only+${sport}+fans+are+%3Cadjective%3E+%3Canimal%3E+%3Canimal_part%3E`
+    //Occupation insults
+    case 2:
+      return `https://insult.mattbas.org/api/insult.json?template=People+who+are+${occupation}s+secretly+eat+%3Canimal%3E+%3Canimal_part%3E`
+    case 3:
+      return `https://insult.mattbas.org/api/insult.json?template=So+you%27re+a+${occupation}%3F+I+didn%27t+know+you+were+%3Carticle+target%3Dadj1%3E+%3Cadjective+id%3Dadj1%3E+%3Camount%3E+of+%3Canimal%3E+%3Canimal_part%3E`
+    //hobby insults
+    case 4:
+      return `https://insult.mattbas.org/api/insult.json?template=People+who+like+${hobby}+really+like+%3Canimal%3E+%3Canimal_part%3E+because+they%27re+%3Cadjective%3E`
+    default:
+      return `https://insult.mattbas.org/api/insult.json?template=You+are+%3Cadjective%3E+if+you+like+${hobby}%2C+you+%3Cadjective%3E+%3Canimal%3E`
+  }
+}
 
 module.exports = router
